Update files and clean up codes.
Rename package listitem -> item.
Rename OrderingCategory -> OrderingCategoryItem.
Rename Card -> CardItem.
Unnecessary code cleanup and sorting.
