+++ app/src/main/java/com/uof/uof_mobile/activity/MovieOrderingActivity.java
@@ -25,10 +25,10 @@ import com.uof.uof_mobile.dialog.BasketDialog;
 import com.uof.uof_mobile.dialog.SelectProductDialog;
 import com.uof.uof_mobile.dialog.SelectSeatDialog;
 import com.uof.uof_mobile.dialog.SelectSetDialog;
-import com.uof.uof_mobile.listitem.BasketItem;
-import com.uof.uof_mobile.listitem.MovieItem;
-import com.uof.uof_mobile.listitem.OrderingProductItem;
-import com.uof.uof_mobile.listitem.OrderingSetItem;
+import com.uof.uof_mobile.item.BasketItem;
+import com.uof.uof_mobile.item.MovieItem;
+import com.uof.uof_mobile.item.OrderingProductItem;
+import com.uof.uof_mobile.item.OrderingSetItem;
 import com.uof.uof_mobile.manager.BasketManager;
 import com.uof.uof_mobile.manager.UsefulFuncManager;
 import com.uof.uof_mobile.other.Global;
