+++ app/src/main/java/com/uof/uof_mobile/adapter/WaitingOrderInfoAdapter.java
@@ -9,7 +9,7 @@ import androidx.appcompat.widget.AppCompatTextView;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.uof.uof_mobile.R;
-import com.uof.uof_mobile.listitem.BasketItem;
+import com.uof.uof_mobile.item.BasketItem;
 import com.uof.uof_mobile.manager.UsefulFuncManager;
 import com.uof.uof_mobile.other.Global;
