+++ app/src/main/java/com/uof/uof_mobile/dialog/SelectSetDialog.java
@@ -23,9 +23,9 @@ import com.google.android.material.chip.ChipGroup;
 import com.google.android.material.textfield.TextInputLayout;
 import com.uof.uof_mobile.R;
 import com.uof.uof_mobile.adapter.SetProductAdapter;
-import com.uof.uof_mobile.listitem.BasketItem;
-import com.uof.uof_mobile.listitem.OrderingProductItem;
-import com.uof.uof_mobile.listitem.OrderingSetItem;
+import com.uof.uof_mobile.item.BasketItem;
+import com.uof.uof_mobile.item.OrderingProductItem;
+import com.uof.uof_mobile.item.OrderingSetItem;
 import com.uof.uof_mobile.manager.UsefulFuncManager;
 import com.uof.uof_mobile.other.Global;
