+++ app/src/main/java/com/uof/uof_mobile/item/OrderingProductItem.java
@@ -1,72 +1,93 @@
-package com.uof.uof_mobile.listitem;
+package com.uof.uof_mobile.item;
+
+import android.graphics.Bitmap;
 
 import com.uof.uof_mobile.manager.UsefulFuncManager;
 import com.uof.uof_mobile.other.Global;
 
-import org.json.JSONArray;
 import org.json.JSONObject;
 
-public class OrderingSetItem extends OrderingProductItem {
-    private JSONArray productList;  // 세트 내 포함 상품 목록
-    private String conf;            // 세트 구성
+public class OrderingProductItem {
+    protected int type;         // 상품 종류(단일상품, 세트상품)
+    protected boolean selected; // 선택 여부
+    protected int count;        // 개수
+    protected String name;      // 상품명
+    protected int price;        // 가격
+    protected String desc;      // 설명
+    protected Bitmap image;     // 이미지
 
-    public OrderingSetItem(JSONObject jsonObject) {
-        this.type = Global.ItemType.SET;
+    public OrderingProductItem() {
+
+    }
+
+    public OrderingProductItem(JSONObject jsonObject) {
+        this.type = Global.ItemType.PRODUCT;
         this.selected = false;
         this.count = 0;
         try {
             this.name = jsonObject.getString("name");
             this.price = jsonObject.getInt("price");
-            this.conf = jsonObject.getString("conf");
             this.desc = jsonObject.getString("desc");
             //this.image = UsefulFuncManager.convertStringToBitmap(jsonObject.getString("image"));
             this.image = UsefulFuncManager.convertStringToBitmap("iVBORw0KGgoAAAANSUhEUgAAAHgAAABLCAIAAAA06NSGAAAAAXNSR0IArs4c6QAAAANzQklUCAgI2+FP4AAAIABJREFUeJx8vGeUZVd2Hrb3Pufce1+uXNVVnaoDGt1ooNHImCEmYKImB2aJFGfRJEWPpCFly5ZtLVnysq0lWyZFeommJdpeIs0hOSQncYABJwIEMMA0Quccq7tyevHGE7Z/3PdevS40fVatt27dfL+zz7fjOfiXf/Zl6jUAQkQAQEQhVL5BRADAzIgohDDOEXVP610FzAwAQghmzDfynSSFEEhSKOEJIQR5UkohFBFJKZ1zAI6ImJ211jjNzDrLjDHGGGttprXWOk3TJEmUUogoiBBRIIleA5REJARJKaWUAABsrbWYN0H5V+RvCAAIAgAYu+/MzNZaYwwiaq2d1cxsTKa1lgI9T3pEzNa5/GogIqDevYUw2oZhmKapEEL6nhCCgIUQ+QndJyIys8z/Z2bnXB/lPnYA0N+fb/fPRETXf37vdr2LoHsHx0wETAAATHx3A8gvZ0REYnDAzI6ZAQilRQeQwyQQ0TlAdMgkJDAQMzsHjp2UjhmcA+ectRYAGKx1FhEJJTKjIGBApP5rAYDrvWvvTcANNJs3zVmWDFUr7Jy1nIsaM2NXwChJkiRO0zTNDxljmFlSVwT7WOegSQACQGYAQGbXP8M5g4gAgpn7Iuycg253IryjMTMAD3ZSd8MhO2TsH3LM7AABGREZ838tg2VwAMSUAyAcWIfESCgVMwIBAzrG/FJ2nHeuZQMgLHe7zLm8J8ACEyIxMBICAgB1RdnlHz8oPTbHktk5x4zOOZNp63QaxcVSoVQqCSGMMdoYdEhEURQlSZKmaS7ayIAWAEEzEjM6R0SSKBdHa63Mn9ftq94Q24ZUjnVfcreB28c93+5ftTVg0fVOdgDE7BDJOUtEALk89r6cyaEFhw66d+jfnIgAGbiLExAKQCC0bJgRHYIgYnYIjMBdlgBmtuwQum/iwCEDcndc9j+hL9f9wQ3dQeOAXb1er9frQRAUi0UhRBzHcRz3vzfntHyU5yMyf1vXe0B+czk49vPWhynv78HjzC5n4a1x1CPxuynJ5qOhd60DdIDICIyOERm7QxhROuecA2cJut1kGRyDY7aADL1tB4DADgWAI8fMDALIgUNCJGQkZwEAebCnBeZ8A7b/mQKQ75akPmMws2O0rtfxSAyUGp2/VieKW50wl7kcRymlFMJBPraQHbOxQghmYAZmzpwFAGJgZnk3rPdo2yh72/Y2QR7cyPuZmQHynRaBBi/h7Q3y13KOwRqwDp1Fx2ANGwvokAgJEBHYMoMD3WU1x4wIDhkx/0JmFoDdS7pc75CJESy7/ksOsvMg4jlF5xySdxh0NbxltsxorSWSAGAtO6dz3T74RfkN8/tvAb2NN/r49pVef6DlZN0HdFvLbz3wPJvfjxnZIRAyY/8cCwwIDMwIDoEJnc2VWi7QCA7Z5frRsXOAzEDAlHcHIgIwsOuNHM7Ry5/XE+lckwMi97qXHfAg0Ntw6V0noKcBAZV1tis6JAkFMjFYdsjMwMiYS5XoK20AQMzfAgDAWeOck4PytU2EB3HvcxAx893Cy3er78FfAGCHQAIA2RESMmNOf9DlTe6LP4NlZ3t9BszMDp1z+ScxA7N13ZPzpxMRWmPIETEwAAvZfxN7F8W7rU8j2PbCg2SdCxMI0TUQWbBjErIvYc6Z/BMQiRCEUIg9BcZEAgdlufuUXBkiAyEiA+bvwZBvAzMhEhK4vJO75iNAd/xj70scgLW2b5lAbz8zEwlECSzZCUCBQISydzUJodjo/EHOGmJnwRmAzDnjWFsLAI4xyTISSmstBFnLRCgEOccMTiGy0X7gsctyTQQAxjIAkBSOHSHBgKVPRGxs/w37+xGRrSNAAMyBytkgN0W6Bnt+spTOAoMFh0QgpZeDrJQQQjHb7uiBLRrMgd+S6G2i3d95lzHnmGmLN7r9NkAjg4OgK6rO5KgjMzrpnEPKmdRZmxT9wFprTEZEWmsAslY7QCml1mmqdZqlnqc6nU73DQGZHViw7IjZOVvyvTCOfN8nKSG32wCYMUs0KSkQWGvo+VY5an067g8mGGj5HiLKwXWOicSAX0Yo2DkkiYgohCAiRBZCCiGYCQZ9C8f9u8nBcTT4SOdyn+0uBmfmvue4RTV3u153Nwtoc0Hpdgkis0DEwPOyLLPWWevYIhIJVMZZXymtI2stWyMQJEIcRwiOkHSWEknnHEplnRUqQCk6Web7BSdkM06llL7va60BQCrJbLU2AE4SAVCu3JhUb0wCMrBzDmwfAQZmsIAO0CExMhNhz7XYoso+KeX9h4hEEpGYjURi7hk5tGUXbAG9DaFB0LHvrQAQ5c/b0pN9We/v6Q8FAGDWgBbRQyJA6BOczqySvs60lJKkyLIEAEymM6OVENo5ZtZpJgRprRFZp5mUkggdYJZlQkltUuMsSqGjuFwue8pHxCTNJIksy3K4PUFSSrZWpxqJhfS0tYg9HzjXBLDFqoMqsStQEomAkJAYuhqqe2hwQPSVqsAt6AZRlV0LFxi77ILA4p5GxSCrAFgABBgYU/f0WYgZGAABHaDhrRdlbVLlCWDrLCOD0RkRSUGOvThLs0wDCaFUo1GXwmu1G0EQdKI4H86+VzDaIqJQjNb5nheHHd/3iQQDGM3gQHnSOWc1szZIQEzIxEgAYMFBbneiQ8wtRgR2OTbIDtkBMwEiEiIgMCETAkOuvnowAgMxEAPk/gEAOAAHnG/n5zMgMzg5cN12o63fvYNENthLg6fjvZiaLZAgAEQQwBKY+sNWKWmMRnKONQFK5UwaAbJNOj6i0W2wjtN479TwrZs39s9Mrq2tDA0VPU9GURSGTas1MxeHhzaWN2rDQ8ViWccd8gObGVKeNeyMQJIEZHuazTl2SSaUdMDY6+98LAJ0TcBBic7Hp3Mu95sAAYG2bCRGgHxY8xYX9YcybsHS5eh3Si6gA4Q8EtQzfajHGHn/AeROCOT6RLwT614jBJ+tdCgRBKJCkMA5v1vrjJKgJOq0g5Aqz5q0dXA6aK0vcdU6Y4MJr1Yzjx975MaZM/dPVc+dOzccDE+PBYc/9/7GlSvGWCRZemLv0sr6ytp6eXx4dXVlZGjkzsIdwyT9IopAqgIJZRiSLEEUQgjSmUBCBCLKfUTHlpk9KbvA92zyrucvABGoa7xAPnSJkB0yIHatvvwKRwIoF/0eAbge5PiVP/7DvsAOcjwi5gZsn/LzfsuV7JaMExJ6iMjEgE5iDq5yPXtQokeUO6tCSiklkUAAI5QRaIhTKWBjfdGZdPfOGTRJa+lGp74yMT4CAGvrdQZbLSh0dnFlWXlBoVwJw5CZNzc3JfDOXXtW1tszO/eMj44tLCxM755d3miutqKDRx62Dm/dXl5Z3XDo+V7JorBUkEIJZgQHubNHhMiSYMsWBnDOWdhSM9bqAQTycb9FmETEAOwcA1A3HAow4Hw654yzzjnxk5//LAmRj5xcvwIyCYIeJSHlhgUD5rFGiYhITuTBB0fAHiMCZ4FPWdQUiIQFEn5mNEhAZySCJ6QklIJ9ASZtlwMI23NZuDBcdKa9hMnaA0f2vPqtr0odHbj/MGm9cPMKOHfo2BPLiysQ1sP6qpAyjuJHPvnZG2fPI4onHnl4eqy8snhHazBZPHf2DYrryzcuNtfnhyfK+x85bNcXSzKdGvbf9a5Hlq+dUehSY1EIcik77ZCE9ASjM1bJPPRq8xgMAhMCEkpEzON/XcEFAhRIApEQcuMpJ24BgAQCQBAK6sbK8+sAAAEJSfzU5z/bH/K9buzbztyX8h6jYd4BCIhdgkMACUiIxlmtFIJlScU01X4p0CYTyIFHyJlA42Fs4rW9O4Y4XYN4bedYKVxfDCBJm6vcXi9JkE6v3plrbayODZXjKHn7zPmnn3ziR99/bnSoYozzguL181dL5VpBUrkgrl44G4VRqTr86Ac/oBqrreV5H6308OCDR/7D7/5OVl85tGfm/NlTp197CeMm2swrlqZ3TLJus9XGWkECgYQQJAyAJXKEjND10gR1w7rYtd62t350XyARUf7bp4RBeuhi+2d//IfbQqBIW6mTPMI7yOs5IxAgEXVDsFAgIkCDaAkdMQH5JPxER0KRUsrauFYQZc+4eLMagO5spu2mBNZxrLX2lPA8CtsNz5dR1JEChRDTO/bevL2gVbC0fGffxJDNEiQ/YvnEx38STHbmxW8LHZFU9Vb40GNPeURX336d406WxtYvXG2bZz/80fqtq0t3bovyqM5sZ31lYmrXYgdr4zNRmtTGJjUEFgvkjzjnfM8Ba+URMAGrbgSKmB06QBrQQIPY5U5GX23m5zAz4KCxgMycx6fuynrA3aGWu/VvX4/aPHTQO57bbRZBGOOQVGqtlAScost8cpxF0iXVwCStxbGyE+m67Sw0Fi8HLi4rrAWy4ivWxvN8IpICiBPQ8bUrV41xJgk/8dM/NTw6rg0DuAeOHrZRdOXCxebaEpqstdk49sADcWv95pWzV65eqrc7mmF9Y+Pn/+l/OXXkEKWdzvqKh/qDn/uEsOHy7Qs/+cW/H3Dr03/3M9fOndg1XpiseS6ue4qJQHoyj9NLlF2JRSml50mvnzMbTJoMQrmt3VP2hRDi85/9TF+iu427BOHcO0Kj3QBxl7W7VhIDABhmEiIzWkjF1rFOPWnT1sZo2Xvo4Mzy7YsPPnCgvXQHsiiQbrhathbiJEYJLCHR5r4Hj54+f3ZkdEQYZwz75VpmQCnvxqUrC4uLteGhQuAtLC/t2HPg+tUL7ZW5kuRCodCqr1468/pQ2SuUK4cefLBQHVre2Lx96sT82z8+8YPvHZzdo11aIb2+fDuMOhOe+uFLP7h+7ux9B/ffvnz10OzeLIkBXEbokIiVAClBEhKTcCQFKIUq10zboNwG690QoxDY8ydzHgYi3Aq4bGt328t3eYm5QTNwpgVwuWUppCeEcGwQjI2bH3j/0/fvHVu4eX5mrLJy80oh8KSUSaodCRWUKAi8YjFzbs/s/o31zd279yjlO5LCLzTDztHjx1ZWNyanZj70y79ugDKb7dkzQ8LcuHxxenra84KlO7cOP/XkQw8+3IkSJrx09VqhNvr4U+8Wztq4/chDD2ZxO22snnj5++sby0g8d+t61GqAiR8+evBvvv+N1TvnFy6/QdFqgJkUAOAsYy8hKAWI/qi9p/z+beK8jZr7p9E9XHDk/t/gTQEAGAEoT+u5Hhnl92awzNZBHqxyBPbQ/unr519vrM8NVxXbzPf95fWN6uhkJgKragl5s4/9xLW5Zc+vLK+uedJfX1pLY23QT1kElSIQf+Snfzaojn3t3//+8sqaY/QknXn1JcyikeEp4ZcmRybOfP9vltvcgbLvV2qVoZH9h944dbFaHT/6wHEA8lB4NnFRe3R0vNEKx8em9s/u21xbPnXilX/7H37ruW/80eL1t8eLjlqLVcqkZ51yRjAjCHaCDaCxqPOU5jYR3PLI3nHIARgHxoFltIwOKN8Qn//sZ7Z1FAx467lhg4MhvTzlBoy5f9rVljmNOEA0JvXQjtf8sq8LIhOo2aZJEjYa9cmJHY1WE0k6EIK8W1ev7Z3dk8SJINHa3CwEHgnvvsefWl9bF4IarVBn+OJLL//0L//iwWeefPM7zwlENtkzTz0x/Nij3/iDP3jgvgNJZqxXfvoXvvDiV/60WAzQ8dyt62DighCXL11QvozS1JCILcug3A5tEBTXV1ebjbUfPP813yfB8sRrPx4fGz175u0Dhw9ZBMueZamAEYDJAeUpHNoOUU8ZvnPE5w7KNiUJAOJzn/l0V4qBu2wyyPrdtAUjQs5C4LAXsQYGJOhFNgiRUAhCsDppzEyU0TRt0ioVVdxpjQ2NTBw+eu2tN2rliodeoJSxWbVaSsJIm1QKtDpzlg8ee+Tm9bnbd1aOHX5ocX6V/NLjjz/y0tf/ZOnkiVq5VAw8k7XC5sbbf/388YceaDU2CqUgKBbt+mqn03zm4x959fmvFWRqk3bYqUtFjST59D/5Z2+dOtsKo0cff1J5ldW1zbDTyuL2yFBpYmK8Nrbr4H2Hz739unCJdemuPfs6JnDsB4KYrfSkAQYkpK55jJTHPwgQSQoUhILymFP/b5CuextCCNG1o/t9dk+C7u3P3b78lvmBPBONSJA7SMAOTPzI0QPVAviUCdBgTV5ocvWtk/fdd1CicpajThgURBS2rDb7jx879dqrtUqlVCwt3p6/cv36R77wBRnFP371lcnpHZMzY1NlmUX14UrNkzBUCgitTpNSseh5fphkcWYXFxenxkfPvPYyubDoq0YYZQxKklP+xmZ4bW6hHHi7ZnYJWTx58vTnP/2phYXbrc016/Dwg09Faba2OAesl5eXduzYUxrZJWSJ2CglEqOV8vvQbLFtD7xtpkW+wb3qjG1MTY6tY9sLhOfJRwuwtc1se4aeY3CAeVBRgKXuVWAAnEQiBo9QQJZE6/X6QmVqotPpOAGp08IPDh65P07TKImV8mZmZ602hcArFYLNq9cEu3Lgp2FUKXiPHj2w+Op3v//CVx98YFa35m+88eKV0ycgCU3cihqN+voGGCiqgk51lJowsQcPPzg2OrFy+85YqThUqtxZWn3/J34uFtWUVavZ2T094wGVPe/Eyy+vzs9fv3y5UqndujkfqHJrM1zfaOzZf2Ct2Uh1pqx+7XvfHgskZh3ylSbhFcpIqm/e9aMI+e82G66PrJSelJ5SvlJ+vpHvEZ/7zKd659xVSnG3YqWt7ZzEmSDPaSMwWgS2hgWRYMsm3LOzBrq9ubBQ8INiMUBGYoqiCNEpzzPGzt++Va1V0kynqUmSZGZ6R7PVKJUKni+zuJPF7WqtLCX7aDltVwpSCnDaBYGvkJIoRkC/ULAMjzz1tL9j6uXnvjlSKQaerLfajzz1E2sdffSh4+dOvzU+Pn7p6s1SoYQ6EYSNZvTP/vW/+cZf/EUUhkoiKf/24srC4mKSxr/4S3//1ZdfLlWGMw179x/shIlBIiEdQ+4DD6IJd7dte3r56C1tmV8lfupzn8od/Nx2BnC58qTu/0yYW9bdPwIm6ClBYu7m+IFACGYJZnw4qBSdQCuRJh46Nn/64mhlCJyRxESO0SVZXK5VM8Mg/TQz+971rvmbV/yiIg+iqO3nqSZy7IwJm9WC3253fN8vFoIk6bjUEAgUiEKghJXFuSsnXhqrBp7yNIjZww9NH3vk+88/f+v61VIp6HRC6QVxFMZp6BDKtZHG+sbJt99WvgpNhIqUEFGnRVKu1xvrrfojTzx548qNarlSG500QKiICQgcvUMZYi+JRfdw0AXe3bp7P/fpT2zZK+gGeyOPkXKeVO/Hprvd1s22O+j6LIQUKGmi9sRoMZAZuoxIZmvruekO5IwzKNha4ykVp7E11mhTKPjNpYVbt64XC761aeCrJEwqlUqz0wKXSnDsHEpZKlWMyXyhTGr8IEAhEJ222jlN5ADR84tesXZrZfPs6fOf+8yn5+ZuguOnn35X4BU+8P5nD9x38MzZcw5wYX5pcmJqs7FZqJa1NWCMrzyUQTOMZSG4du1GGpsjR44F5RGNkomtyfK0EA+kGf//G/Zck23etcwrP/JcwzuNmEFliCAQUYAAdMgGyDIqQAJCAEcOszSsFFSlEHgS2Nk0yRyiPxQYZ02WkkAEStPUl0ah0cw2yxxgq765b+c0ARtjozSSntdOw1IhiMI0srpaKpnUxJn2KYjDTrU2orVuNxvV4XK0uVkdHU4AiFSWkonNwSMPV4ZHzp69NDY0fmdpfnJy5tyZSzev3wRFpUrValcsqEJQPHDw/ttLt0n4vlCZdiQkAgd+IdOqUh7PDFlWyCiB84oduBsNuFeOdACxvt096NOBrFUKvXMZ8v7gu7C+m68FAAEYxLwQjpCUI0REweAJX7fr1rJSXpKmTEoFykEG6FLLPgmWEkTACCgEm7RSK2dph8EoAVmmjTFhFI6NljKdgHGFYjXORCNMPekbUGCgUB5JtE4yW6gMWQejEzORTr2gllknA88PSlrrOIwIMIqSXTv3fud736+Vqkr6LIlC6ZyLI90Mw2K5qvyic04AlUt+pFmRYkGkMEx0ZXikWqkVUBhOPE9am6c2iAkH8R3E555kjXebcLJUUv2uyIPX/QB/Dv5d3QjkSBCQAkJ2AgQiOhIspLWWkIujYyljxuAXK45jBLLWCiFLhQIyJGnolYbRCrYZYxJzahhU4LN0aRo7j3xRtGiLxWIcaUcoEEgEYSsJqgWv4BmdMSm/UJJSpmlsrEXp+UoSO4ekgkKnvV5fXxIgSqWKcTg1s9uaDKRCh6Lib7SbouApVMbagqpm1hhwLCQJ8v1SnKbKl/VWNDI97hcLHgglfWbLKBiBQDDeBWU3xUi5FNI7Ud5GDLJY8Pq7BrsrvxHxQB4MgEk4JACnAAUbAgFAjoRFCAp+lmWIUoPTjgLPh9RIRYrzDKwmQUJ7zoIiqlVHW52mZq8UeCgEApeqASrZarVIKktSVUqMEIhafaNeGqn55VoYtYqVmjba87wwDofGp8IwFoylaqnVaUdZPDw+3mm1nLFJ7IKKB9JLjS4Hns20SdLNRqs8pIy1qXXCgPSrcZRacizBJcYwGxax1rv3H1pe35zeOTE8MhR22spTgwLLuBV9y20ul3t1KO4p0YPQ4Vuv/FXfcMm1X09XynsODSYBbCUYQhYgGIVGdsAW2PM8m2Ql32u31jc3FoaGgsXbNyZqtdFaxZOghMxSnn7g8JUfvTSxYzpzCj2vHW1USuVova2EJ3xynDlnSKD0PW2zdtQGJgkFgVKgZjBB4Dk2aRoTiIJXZaZM21KpIKXL4khbJBLGuHaYslAWUJDSaZxmHZ0mvvAanXZGkGSGQwvGGsepM+00Xt2sO+f9xHs/PDVzX2xocWXj73z840kSKyGhhwnAXdF5ppxGBgzfd7h7gzpPKqX6+QLu1k1JRDTGbGccJiAEFtA1YRhAAgmBwGgVIgKWyxUdx+XSyOyePRuthenpKdPqnD/1xubafM0rnHrl1Oc+8cnXXv3+xPhkrTZTHhnFoqzj6vBwba1+hzwrJAdBECWuMrRjbWNzaLjkkdBtXaoWFEi/qMqTI2++9srs7Cw6RINGs2DpUiBHoCWxUMLPsrBSqGYWnOOok5CU1dJQI13vtGOB5GwENonDlhRBqunO8nojyY48+OC+A0dABFESr693nnjyXVprIaRxjgDEO1AGuEfIkwfKCLZJNzPjuTe+h3h3fIMJ8/TZIMr5AGFEVIiMmCIygwQSjhCRrDa+ktKZgpLlYsW6TLsOcJZ0wo3Fuc7KzcbNW9580llYXVi8JQGvn54PNaQVWjBu+jh8+LNPV4rp0uLtYjC62fQ60Q6Aoq+yqsRxr1RCGbU6I2Ojwzsn62H7sWc/kKyuBn7hzrVru44dTzY2g907IYpAeJBlOomlpDRJhBAG7Nyd22NjE2sbG4njVrh2+tT31tdvVyrD1+dWSqOzO/Y8uGP3ERmUhodrni9Jeu9534cy7Tv2iMhaPVACR+9E8B7I8tY8oLtOuHzqpa6Qb9WN57cdtGlEz8IjIgngAC2gcywQiEkgMnFWVLZdXzj19onP/swv6lZns73ZajXCqNnZWOLWytxbJ9d+fP347H2H9u/53vMvrN5phDHcjmHno6P/8lv/K1TasHYtrjc8Ofryq3O//OtfswBPPTx8+3x9tw9VhoDBARRGi8ubkRfQ6HCttVafGCvNrYSjU0PgBaVqZW5urlQqTE7uCBsNBXZ1dS0x0MyAJXARnv30R98+/aP7ZkvHju0pFCmzfPrScmFo3/DEkcnp2Tu3rj54+NDhI4eUV67sPGITyKxEofKk7bbKjUGgt4l5bgQPHupG8i6ffHmwxyCPOfesju6tsddLBIgW0HVLYVgA5BGtuKjaaXyzvnFh/56ZLMUkoXZkW1E4OlkZKimRtJeuXO/MN6ETLd242Ky34qwQlMdTC/64f/QD9x95ZHZjeTELswvnF/7kKz/4/sv2t37nny7cuLx85fLaqcv7qpPYMQVZXNpsogqOHj1y+8bVAqTFYnG5ky5u1pXy4ywNORsfHxMpTw6NtFZWO2EzY6juGeKJ0c/8yi+livyKKAbpzSs/GqpuEsZRFvz4zetjUw+devvyUKEwUi4/8cijzUjPb0a/+T/9TpQVQVbA6H4WcBDje7IEIua1cvc4dPnky0SU682ubyNo+6l9H5TAQgrIwEKih0DIhEwSmgLubKy+sXsG7sxdASp4/ujyRvrUBz8KcWtjfWnIJ5cknKHnefOn31xcWLm+0LEYJFF6/c614qh6+0xjuAobdahVYH4Z/o//+/+6fu1m3Nps3lmYe/2sXm3OX9oAAA3gARQ9NTZUDjfrHQNWQGih4CN7EAG3I/AZfuYTn3hgz+xf/Okfj88M32otf+xXfkHt3FGenKqO1hobi4HXuHn52zsmi4iF63Pr33rhzbHazNGDD9aX65trm+04rUzv/pe//fs2mMqcLwH7Nu42oN/ps2CvGOadIi9+44u/lnfV1n1wS11yt/6sV0FCCEgMgCwJkRiRUTgpIVO4PlTN6ms3ahVPKRkU/STLlhfmp2b2nT917sQbr9eGqpWRoX/1P/z3l2/dWu2EUCi204gKst5seM47fHCH5WTnnumL15p/+bWvd8KOIuYsTtrtwCv++M3zLQ0JQAYQA3SsWw2TuoMQoM0wtKNy/NmjcocXFhN/pvjRn/vsn7/wzRNvnWibaL5VL0yJK0tX6knjwOEjRGqkOvrS9146d/JcmsjFpfDU6euF8tjY2J71FfOf/uhHmSvvO/L0Qj18+n0fKlfG2SEB9iOf3Jv58LcpPQBw7q5y560zL739yuA1uUR3YQWArRg2AoBD7k6FdCwQCVAyCesr0fDUFZNcr5aty9LEUCd24A1fvXZnY7U9MTa0unz10P37NjfqJ378Y98T7PDoA8ewnRKHAAAccUlEQVTq9frSwsKd6zfv332ftTYYLl+4ufyb//y30hSj9npYX1q5fYWM/cP/+KeVgqgWC9VyRZEK250kDolBa+6EOtO82Qbrwfs+NpswNEK9ttJZm2/srY17AM1448n3PLEe1zXJzRDuf+CRWm34W1//6uRIcXHhJhNXhkfGJmbGJ2Z/7999dWxobHW145fK7/3Ye37vD/+fKMxQeI4HNNtA6uSdFkjflNg6f/D38smXt8S+Z34TbZmHW1Z2PocwL+XvztlzxCTA90Sj1f4RutVTJ05GLZvpImP57KVbqeZnnnzyhb/6009/+icuXTi1a+eBRx555OLFk61W3erU6QTSLFByfWXBWJs6c/Tp93/qF/+LxZUGRpu3r75RFBsLty6YNNu5c6eU0hrnDBOAEqSQENEyrqxvqEAx6dWN9dHRGd+r/tU3fmgNXrvJrQiGh+Dow3s1sVcoLK83lzfqGuXY2IRv4dbVG5VaRRXLGv1z527dv+8h6byrl69J5Y/OTD33Ny8xY+ZYW0e9crhBcezrsG1WM7u76aEP9LUzP+rvHQR6sOu64gzcfwDldXzgEIQAT4mOUjfn58698crZH3zntWff98nnXnjp6EOPv+c9P3H57MkkXF1cOH/zxo2oBVkGz374wakdY/tndykhPIDrV85WygYpq4yOHH3sPVzdZR0v3bwQQD0L50aHKMuyOE0z6zzP87yiEqSAQWtrjNHOcO47IKEQKtAJF1RVeGX2i5mDi5euLS9vqtLoqTOX19YjUEFpuOaMk5rOnb60WYf9h3dev7OyMK93TZXGamP79s7+9Xde/MinP/XYu971y7/5G51WUwjF3VLGbt0M9Kr07wl0d5rwOyc/XDn78jb4cwNlmxm4rRERdef0EXJAlAlYGhkq/vgHb2cdlyTJxYsXK0PDTzzxCGbR66+8uG//7nIlOHXyTefc8PCk8NSVSxfmFhadU+Wyfuyo/8QTBw69572ttXqpWE2zaOHW+UoJKyX0pXAgLXPqdGK01pqcVaAhM6aTprEB8IyjSq1WqlVJilJ1CAol6IRJa7PZqjc6TZBDUTbuYOp3/ve/eOvtG1MTAaFzWba2CuWav7Kavvf9T586dbLoi3Y7PH788ddPnQM/uLS8FIdNJHIOuk7y3UBz1/yld/qBfZQHsRb/8Ne/MAhrzhbvRLnPId1tEEQCMLfuCJAZIQnd//a//P7RIw8d2DPz6isvCnA7JsZ8JUqVcrE4cvrtcysLi7dvzocdOLj/aG1o7Gd/7pdOXbj59W+eHBvn2sjI3LX5qB2jNTbteBI9CcoDUMowORkUq9VCpVyqVIPAN8aEcZIZQlX2ikPTO2eLQ6MoVXGoBuiay0vzd65HYUMJMoxRSN/85iv/4l98JwybYYePP3S0sV6fGB6vVYq7pmdLQeHOzbmp0TGddCrlYHlzM0z17oP3P/vsB8uVMnSLiHIcuDe7rls1hIiDbnZfeN/pNCKi+NIXvzBQVtPNEBBhPoWvW1qZey95BRO4PDUFiAyOgS0aBuNJpYR68tFH/8/f/bcFP/nYh59560evD9dqX/urr5drQ5tr7b/88788sG/2v/tv/vnF89e/+93vP3D04Le/89z7n3nmxydeurMa79137Mxrl26cv710Z+XMyUuNDb1v9pDFtB23UgekPGuFACGsddpIUShVxv3KJAbDhcq4FX5sneepxvJqe201a9UlkF8cbiWFF1+8+nu/92a7nh45vLdaHR4aKUdZeuzYwz759bWNpfnbI9XScKUa+KLTacZZ8uCjxzTi+z70yU/+5M9mYUqYl/PnE58BELuT2PN57NyvEvhbXfO+aIovffE/IxK9RBUAAHK+0gAhAjBBd8oA534PIDkGdsQMLMAhW3QAYLUTQNMTk6OjFSWyidHR1mbnIx/+6Ne+9Y1OnBlNG5vNjc2NH3zvxfm55ccee6wdtj1ffe2r33jo+PFWpFtNvXR95RMf/sSbb51E8MM23pxb2rl7V6FS9YKq8orFQpmNQ3a+73lekVGh9MvlGlhAx05nnWY9aWzoKBTOIQadOPiv/9uvLq80D+w7BK7UamdeEIRRtLa6Njw0/Nabbx479tCuXdN35m5GcVhvNBhg17495VqlHSef+NzP3XfoiHXWOE1EiN0BDVvhIHLW9egEALYyUNvr63pYi9/80q8DIAOSkFIoISSSYAAhPSSyDkhIxyw9j0gIpRyiFFIIiYSOwCERCSQpAT3PS5JYSJVq9zu/+/t3bi3VhkaXllf3HbhPk/eZn/qpdtg5dOh+NixQrK6tA0mvVLl5a25298641d45ve/ylZsLK4ubjfYLfz2/c2byD//TczPT+w/d9wBY9pTMko5O0zRNESCNQ5dFWdjAKOSwo1tNTENP2sATqUahxl544dzS0rpSQ/VWOja+o1wdvnLlRnOzBZZnd+1aXJnfvW/X26dOLqwm+/ZPjU9Mzs2vl0pDh+7b14nTL/zn/wgKfuoi4eXTnwUDWsfdTCuQc4xS5GaYEKKfJ+yz6yBl5x0gfvMf/3o+q45IUK/mOq/5IBKe5wkhvMAnIhKEiCQEChKAREhSkVSCSBIpoZI4KZVKnU7z3OmT+2f3PXr8+NkzZ2b37w0KxaBUm19Y+PrXvvLMu55S6F555aV2GAalysXLl4uFwsbKYtQO9+zcNzw6sba+trLS+LVf/aXvfPfF/fvve+HbL77+2onlpbXvPv/dg/sOFIpBFLZ1EgInadSOWw0XRmjSQKDVUSdqJZpRjD737Teee/46qcLY5E7LYm19bXbfvuWFlUqpJgCWV5YnJsakkrfvLB44uHN0ePjixQtDw0O798ws3Ln6d3/h7x1//4dTbQCMcwahuwTFlrGcr7Bwd6aqvy7FPVMwiCh+4x/9CiIQdHPevaIkAGBrjXM2TRN21mgNCMwOBSNbdMyYr4MB6CyxRSRnGYmHK8Vdk9Vzb78+u3tG6xQFeJ6/ubH5lT/5o6efOFoKOGptCIWyVOjE4cTIBFi7Y8fE7O7dpVJhY3MdobBzZv9bb78NyA88ePjh48cmx/ZEHS4XJv78z775mU9+vFz2BEfORWmYSEAlUOs0TJqJSUj5sfbjeOw73z+378Dh9UaYWXTghofL9c2NkdpYFCaLK6t+QZWCwuLtpfHxqYmRsXOnTk9N1nbvqrUacx/7yHtanfbDz3wki7KiJMmYz3W+iwoAYWC6X3ffoMc36BD2mvgnX/oHiIxAgI45T4QjoHOWAR2hQGIhZL6f2VpnnDPsXD4x2zqD1jAba0Aq6ZwJ22syazzznifu3LxG6PbO7rx46ezH/s6z733mkcDLxkeD4aHC+vpSuVaM03Dvrh2lQD35+KO79+x85fWXP/XJTxLQ4uL8u9/9xMrK/MmTp6qVUqUyGkf2ysVbqyubzz3/4mc/80FZliaJfU8RQBanUglLLskcUM3h+D/9r758+MHjS2vthx5+7PqNufsOHqjVypvra2E7VFLu2Dk5MTF63+yu8ZGy03EcbZRLODHp7Z+tffhD78oyvd6In/zw513GJk3zeOX25GxuewwsQZA70IMob7PzAED86i//fKYTnWVZlhhjjMmMMVqn1lrnjLGG2RpjnDNaa8cG2CGzAO4mZxkkWAFoGS2zEAQm9PSmC9d2Tk/s2TvtFcS733O8Gd1sNq+PVuQPv/etTmdtcnqoVHTWNM5fOhmUcOH2jaff9eTu/bvW1pcunH794P6p//ePvlkqxfUNp7Omsbxr1y6ppO9Tuez9j//z83/v5z8UFApxFmqdCFIAlGqjgrFS6f4v/eM/8IrYDuHoQ4/Hqbty+QICTI1Nbq5tmiQK280sy/bMTD7/zVf37jbVsi0V4vc9++CevaVSKQbSa6tJmBXf/bGfThKnPMUIDLRlUzBCL7fXIw5i7vmEzIOODGybTHfp7It57OKeXG6tFUJYa6WUee0wABOgYABGix4ASGeA0KBCqRynVRHHC6c91wImY0ylNuQoVYU0TZpFGbCxjXYjSbJadSxJeXLXHigNN2+s/PCHP7yzPP/y3/z4+OFd9dX19z77/qXlNYYig7p06Wa91aw3NnbunNp/YO/EeOlf/5vvfflPfubgfWNxuMZploSRQ/S8yX/329944a+b9z1QOnTkGJL35S+/+J5nZmuV6ki1Mj93zWbtZjMcmxoDTn7+8x8KlFnfXHWcNcJGUCDrknJ15ocv3v7oZ//hsz/5D1paKuUsWxyM0vEWOK672Z+phoMW9DsdPbx85sU+zQ+6lTnEeePecguM+fogTAzk0CEBE7J1QCgDJrQuqcnEbFyDaG18dKS+thp1YuWBKNjxPaNRe1UqTuPI94vSGyOv6owjocAC6Bjsxu2bl7I4WVtbS1KzuRllptgJXbE8Vq7VlpeX/EKgM5clqefLuVsXjhzd+fhjh4qUKMJmy5w+c+uVV6+MTewLU3356pU9ByZ37pyur0XnT11+4tHDe3ePjY74jhPDdmykHJBpbq4AQL3RQuWl2pDntTvej04s/fs/eM7bcSSyMjUpqdxf6EXseGtach/obfbcYHjvLo/vytmX80KD/i8zMlshFICT0nPOCKHyXwYw7IgBwfXmGwjIp/5Kj9myTYoy4/aSbi4Jm7q0U1K+0VHk2uBlEzOVVmdNkkby2JUd+1IROgYDneY6pHXgLOp05hdX1tY3h4ZnHnn0fSDKxfJIuxP6BeWcS2Lb6SS18nCztWlce3nhStnXNk026lprT6gaUqHebAclGWXrAiFqQaAC52J0UZZtktSJSaSA8Vpt59RkEoWpztqRceQpv/rqiQvajv/2114MlzIsjXLgO2Zg2wd6cOas7c7a7MaF+vb1tmB0X3Dx8ulX8unkgI4dIjGCIAH97f5RBMEIDsgY40lAcIKZpLKgWAjnnEBU5CSnPpnNhRsqa0DWKRI0Goso00a0vu/gbs0dT8bWpVkGhUIpTULnnNMsWKnMO3/24ma9WapUZ3btKZWH252sWBqOE+0VgiyLHWoCTJKsoIqtZrMQlLI0crodhrGzMooNqWKWGetckkaIztgs7qQAlMYJg/G9THPiGIxxLrVskocfPug4CwqldqRvLzaXN6OxiV3Te4++90M/E9pieWqvRuHJbgVpzg/WWnYghABBzjljjO/7/ZUj8iU+BgV8i0yunXttUEX2De9t3dLtHBIOcspGiRatZqQUFKPwpQI2CrTMiyTjRrRyoyQS29mwpmVdFCZ16eH+B/e2m7cYEl96aRJLwVE7ki546/VzC9dWx0ZmqsMjQnrFcrVSq62tb5Yqw2EYoiCpwDmD5LIkVTJwGqQIWq0WOxOnCThlGDpRmGukNE2dBSlEp9Ox1hIFJJyxIZKT0kOQxrgsDQOl7z+8zzgrvMqJt260Ejs8OlSqjRSL409/4DO6uGNsZn9QLCmlnHNaayLyPC+fEJfPrpVSMrNzIKU0xrzTLexjLb70xV8jEkLIrs9CAoUgIUnIPF2YuzOIxAzWOscUBAVjtZQC2WpjvdKINowEJok9JMdg2Mu0W1qYrwQq8ACdzuJEoTBZWixKrQCkyNLUlzJrp3E9+4+/9Z0DU9OH9hyMW6FOTLVU7TRbaRi1N+udRpOzNOm0o2bdxAlop6N0bXEZHUeddthuapNmWRJ2ojQOARJrQpMk6Nik1mqT81uaZo7BOSASkkVzs66Up3W6cLu5Y2qcVOHCpdtr69xpWURSisglYdSSpaEw7S4K6fueUlIICQCOHbNzDEqp3OcWgqw1vu8xu34kY3vQ48aFN6BfAJaLNiEiZlkmpRTYXf5NSlkoFAK/EBQr7TB0YIGN5IykSlwBiNilJSUUW2DZCHXYady+/OZLz39lZtR75rEHS55yupOY+uLm3GPvf5QhwjROmg1MzFsvv1WR4zMTuxbuzDsH1hEzS1865xDERn0TkarDQ+2w7ft+HMeZ1hMTk4uLy54nScl2pyNUQWtGxNS0rLVsoRDUdAatZgcApKdSa6WnUpOhY5fparnSikMSwmrbbHdA4vJKh7Gk/MLoeHF4SA7XJBaH7n/3z3ojs+XaRG1orFwu91VfFxDPT5IkX4dFKQUA1lqltlZd2Qa3+NIXf3W7XYLAzHl3OWsRUSllra3X60uLS2fPnLl161a9WR8ZHRmemFR+UUmfGXwp0TlrDCMJvyiFQnLV0ZGrc0tf/cYLrGF23775laWpnTtuXr+xd+8+KYq6lV146/LmSnPnrp2tKCThEfmCZZYkVic6iwPfA8dFv7i2uhr4ARubh3XiOAkCXzsDAAgqS61OuViqRlHieR6i7HRCJAgKRc/3GdARW3CJSbzAU8LrtCNSXjuyG3XXCHFxvaOKFanUxMRQqYLlCpTKmSp4hdFdlZGp2vCk7xd71dAIwIgkpTTWeJ6XO9LG2Ha77XkeM+fE+c4/vHHhjUGUuxnY/gISjq21Wut8ni0wxmGy2ag3os71G1frq+tjY2P7Dz5w7OHjhw4dAIEcdVCqODIWGDhsNtaX7yxcu3Dm7I9+cPPSqYmpcm0o+OTHPmhN+PjxQ6u3b3VW6p16O87imZldzc1Wu9nhNJXAiNDptCqVirW21Ql3TE0vrqySEKVSaXF1EYm8wM+yzDhA8J1DJM84p22WZHGxFCRJojNGkKlGbY0Flp7oJGmWZTrSSWw6SVZvpYCloDZcrJSLZbFnR0WJqDZEQ1UaqVKKxWDq3Uce/zjLEb84nNu4QRAIIdbWNhYWFpIsc84FQSClrNWGarVaLr/50i3vpGm8fv7ElhWCA3APGii9+mtkyK2SlLseUhImiwvLN27cPHP+9NTE2O6duw7s23//kQdGd045kzBhFqc6jBpLt1dv37h15fLJt9+8de16p74wUtYPP7C3HJSGyrVSqVIuFaQA1nGpKJIo1gk45+KwWakWwLkwDD2/nCY2TdNSrZTqKE6TYqHcaHSk8IlkJ0kbnfbI6ES91SSJWZYJ8rOUw9iGUbq6vpYZZyyEsRaeQuGXK2MqqIzumPQrtZGJiZkd5fFqq7F2cdeOUqBc1ukE5emOOvDUx38JgvHF+bVbt26trq42Gi1r7fT09MGDB2vDo8ViMZfRLMsAINeZ915aCgCvnn09HxfMbNkx3zVfow901xoBZGNBCI0MKPKVngTILMuYXBi2o064cGf+1u255dWl6T27hkZGnnjiqQcOHwEdc5KZKNGdTqu+0dm4/tr3vrxy59z62lqrDcurYB0QgnNQrYCvwFM0VBkqFkSp5DubVatDSWy1tqMj4+sbq7XhsjEm1QzWNetr5XJxo1G3RKnm1Nh2mFkLnQ4AQ5xCuVLyS0U/KBdK1UK5UqiWK7XR4eEdY+OTheFyqVrxS5WRIVX1NuZvvlkt09R4rb5cP3dpPlUzOw8+trAZj07MTE/vHBsbKxQK/bFuHOQLPuaGhxAih7un7ShPCPTUI3eV4V3YD2R8+9Og+0OAgEiI/vrLAmX+bGttv9i9a2PpLMuy8xfOLS0t7d2/d2p8anxk8uCB3eCHF//6j0fUGqSraVyP0qRjvUZUaidTfmFqdXmp1Wg0G+12q9Fs1uMkyrIsTVNwYIxBi44NSmGMEUIotGUvcZq9AjCA8n2/WNEWVKE6Pr4rKI2USqNBuRaUy4VSuVCqFCvlYqWo/MBThXK57BWCaq1WKJXJU9fOvfHSi98JPJEl6f2HDu3de1+xNqmCEkjfAefr795FBfcqwuv563c5gN3lEq6fPwHbjOUBj34b0ADADgZDIr3F4dygs4+95ROstcqTqc6Mc50oajeiM2+95uKF2elgNIgmhqBWERN796zXO82stv/YT5u4qJPUGJMvmG2tTtNUa42IcRjlqsJkGTPnxa5oUx0tz926WCwWhkZHrLWFYqVQGQIIpqb3WPAAAlRKKM8BKF8GxQJKUanVnHGnT5++detWq9VS0tda33//kT179pRKZc/zrLVJkngqcGws3+WAQC9UtM1kHoDIwb0aXj37+paTwogD2ZqtW+BWB+YBKuyVVlJvieS+yzMYkHXARJQaTURsLAECOyXY6IRN1G5trK7PzS/OyYI/NDQd+Dv37ntg9969qBQw6yTJb+tMNugLOGOZu9anEiQkKiWtS7MsEZIQ0WjLjM5SoVAk4QOAKHhAsLpw5+Lly/V6Y3V1nR0eO3Zsdna2VCppbf3/r7Jr6Y3jhsF8SJrZtfPwLUBip00D5BAYSI5tf3n6E9oCAdwGaIM2cQ8xECQ1kmK365FEsgfOy7vrAuFpsdBj5pOGIkXqU9PknNu2/fTps5tYTlacc+Y4LW4jOHuB3prjW774BPRoQs+nKuyQfcDg9fvvrf7mk9qbEoOUkkk1VWYSkSoam6UCV61dXccEWnO3yZrx8vKfV7+/vn336OjoyInrnj59+uibR7WUEELXdaC62WxiSE3TmFnOtSqEFM0EUMCEGRftAhBB+ezly3fv3q3X6w8f3odIT548Pj4+vnXrbtssU0pXV1duSjFH30HLOceYHF8iUtWUks/o+XvtwjrXB7sh8L7MH7/8ONYc6tPM8Ng+2z+f3fOBGcpPh7/6vtXI2zct9d+QGEMsalmBUqNaOCCLgGiLbVdUiTalrtdrVc05v3379vz8vNZ6fP/B8cOT77/9rjk4KJsrp9VfLJeKpGYhkEH96/zN69e/vb94f3FxsWxvPX/+/OGDk8VicXh4oFY3m3VoUslCIRJR13X+hH7bgC9rvlMRQ3LWf1Ul3oZvd/LOgVa5Aeg/f/1pXgERp2xoRNc483a1J+Hutw3HXse6fbGB+TMiuT5lRtPOUCGSIRZB4lQV0KRhZHN+RapERcyvRfB1HBF9dq9Wq7Ozs9Vq1bbLZ8+enZ6evnjxg6h2tVxe/h0TP/76q+OT+3duHyFyoFirlCxmRgxm6kyWhv3J6oODA2+2lJJS8tXV3byS6zgABtPBwC1kYV+W/8BTuJPa4UAPYF079DJDcBaInCn7cfH0L8DMhvNJU0aZE7wiYtXChBQMUESEoSFsyFKpOUQlwq7LqWmsJ31GhyOlxBxHy0nAiChw+vjxIyLeu3cvoDO/i1rNkkXECX0VEMwpQ7GoAPT2Y+Lo3MA55xhjKWUE1Cd127abzYYGznOnjNwFGvel7e6F2CVslRuLzj+H+f/EE5dK36hNHtF2L4QcST1wDExMIpVQU4iS0QBUgKgBrKKFU1NAEVDB2IBiiIDV1GqlGFxHhxCQSKrcvnuniWm1WoXQ8JBJpAhgJGoUGBRUQVEJAiAiBkBVxK6WgOSYllLc5WPmqy4vFgu/6+Pw8DDn7Npj/qbzkODW1vMc4r1Y45tXP5vZqC52B2f2v46rba+PVN27sYmUl6fcSyNgqZgVFQoGZAJWK0xqZoxLg8YsCRhwZ6yGCqCsBACBWVTRCJlAbeStACAxMxEKwUSQA4zXFKDOTSu7zm82vZruN9dcJfobjUzDu2hMCc1fKGHX9oZ9Bvko81DNnEt8eDIf/D5JDcwAgYwUCDUhGisTkCIYkhkoqQ5pC0NmFSEiWBhQ86soyBTUM7T6tYH6XlCHihPKZGCm6CQBw8al67/9Vu7NchMOXyr/RwJ7g/h47jn8BYNPOOhuQ0U2RAhaWqRAWNQJFHqVUwGFABUIjdGcvhzBPK4DTu5JhqYAOlJ+AhiRkYIQgllRVLr2OISDE0wehTIGT3VDHS8PmQRtTGAcxNkGPKT9pUDvH8r/AIUhD/MZjmPgAAAAAElFTkSuQmCC");
-            this.productList = jsonObject.getJSONArray("category_list");
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
 
-    public JSONArray getProductList() {
-        return productList;
+    public String getName() {
+        return name;
     }
 
-    public void setProductList(JSONArray productList) {
-        this.productList = productList;
+    public void setName(String name) {
+        this.name = name;
     }
 
-    public String getConf() {
-        return conf;
+    public int getPrice() {
+        return price;
     }
 
-    public void setConf(String conf) {
-        this.conf = conf;
+    public void setPrice(int price) {
+        this.price = price;
     }
 
-    public static class LobbyListViewItem {
-        private int ordernum; // 주문번호
-        private JSONArray menulist; // ex) [{name : "productname",count : 3}]
+    public String getDesc() {
+        return desc;
+    }
 
-        public LobbyListViewItem(int ordernum, JSONArray menulist) {
-            this.ordernum = ordernum;
-            this.menulist = menulist;
-        }
+    public void setDesc(String desc) {
+        this.desc = desc;
+    }
 
-        public int getOrdernum() {
-            return ordernum;
-        }
+    public Bitmap getImage() {
+        return image;
+    }
 
-        public void setOrdernum(int ordernum) {
-            this.ordernum = ordernum;
-        }
+    public void setImage(Bitmap image) {
+        this.image = image;
+    }
 
-        public JSONArray getMenulist() {
-            return menulist;
-        }
+    public int getCount() {
+        return count;
+    }
 
-        public void setMenulist(JSONArray menulist) {
-            this.menulist = menulist;
-        }
+    public void setCount(int count) {
+        this.count = count;
+    }
 
+    public boolean getSelected() {
+        return selected;
+    }
+
+    public void setSelected(boolean selected) {
+        this.selected = selected;
+    }
+
+    public int getType() {
+        return type;
+    }
+
+    public void setType(int type) {
+        this.type = type;
     }
 }
