+++ app/src/main/java/com/uof/uof_mobile/item/WaitingOrderItem.java
@@ -1,4 +1,4 @@
-package com.uof.uof_mobile.listitem;
+package com.uof.uof_mobile.item;
 
 import java.util.ArrayList;
