+++ app/src/main/java/com/uof/uof_mobile/other/Global.java
@@ -12,13 +12,13 @@ public class Global {
     public static BasketManager basketManager;
     public static ArrayList<AppCompatActivity> activities = new ArrayList<>();
 
-    public static class Notification{
+    public static class Notification {
         public static String CHANNEL_ID = "UOF_MOBILE";
         public static String CHANNEL_NAME = "UOF_MOBILE";
         public static String GROUP_ID = "com.uof.uof_mobile";
     }
 
-    public static class Firebase{
+    public static class Firebase {
         public static String FCM_TOKEN = "";
     }
 
@@ -46,7 +46,7 @@ public class Global {
 
         public static final String SQL_CREATE_TABLE = "CREATE TABLE " + SQLite.TB_ORDER_LIST + " ("
                 + CL_ORDER_NUMBER + " TEXT PRIMARY KEY,"
-                 + CL_ORDER_ID + " TEXT, "
+                + CL_ORDER_ID + " TEXT, "
                 + CL_ORDER_COMPANY + " TEXT, "
                 + CL_ORDER_TIME + " TEXT, "
                 + CL_ORDER_INFO + " TEXT, "
