+++ app/src/main/res/layout/dialog_selectset.xml
@@ -207,7 +207,7 @@
                     android:id="@+id/rv_dlgselectset_productlist"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
-                    android:descendantFocusability="blocksDescendants"/>
+                    android:descendantFocusability="blocksDescendants" />
 
             </androidx.appcompat.widget.LinearLayoutCompat>
